refactor(): cleanup and restructure project layout
updated file structure for clarity. no logic was modified.
