+++ src/core/mod.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod utils;
+pub mod server;
+pub mod resources;
