+++ src/core/server/mod.rs
@@ -1,0 +1,5 @@
+pub mod filetype;
+pub mod http;
+pub mod server;
+pub mod request;
+pub mod response;
