+++ src/lib.rs
@@ -1,20 +1,10 @@
-use crate::config::Config;
-use crate::logger::Logger;
-use crate::server::Server;
-use crate::templates::{Templates, TemplatesPage};
-use std::collections::HashMap;
+pub mod core;
 
-pub mod colorful;
-pub mod config;
-pub mod filetype;
-pub mod http;
-pub mod keyval;
-pub mod logger;
-pub mod request;
-pub mod response;
-pub mod server;
-pub mod templates;
-pub mod utils;
+use std::collections::HashMap;
+use crate::core::config::config::Config;
+use crate::core::server::server::Server;
+use crate::core::resources::templates::{Templates, TemplatesPage};
+use crate::core::utils::logger::Logger;
 
 pub struct Katana {
     pub config: Config,
